requirements.txt
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# redis==5.0.1
# opentelemetry-api==1.21.0
# opentelemetry-sdk==1.21.0
# opentelemetry-exporter-otlp==1.21.0
# opentelemetry-instrumentation-fastapi==0.42b0
# opentelemetry-instrumentation-redis==0.42b0
# opentelemetry-instrumentation-logging==0.42b0
# pydantic==2.5.0
# python-multipart==0.0.6
fastapi[standard]==0.115.6
unicorn==2.1.1
redis==6.1.0
python-json-logger


opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-exporter-prometheus
